Finished testing, noticed problem
The 5 test cases are now written. Unfortunately, there is an issue with the myfree function where it seems that it is not truly freeing the metadata associated with the memory that has been allocated. This means I have to write my test cases in a certain way not to expose this. Hopefully we can solve that by today.